A few tweaks to Jim's changes...
Removed the Game Data indicator from Main (part of the challenge is to read the Game Data in the code).
Gave credit to contributers in the Contributing.md file instead of in the code.
Moved the Builds folder under the 2020 Control Panel folder and disabled the auto incrementing of the EXE's version number.